% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.sim.setup.R
\name{check.sim.setup}
\alias{check.sim.setup}
\title{check.sim.setup}
\usage{
check.sim.setup(simulation)
}
\arguments{
\item{simulation}{A simulation object created by a call to \code{make.simulation}}
}
\value{
a invisible copy of the simulation
}
\description{
A function which allows the user to check the simulation setup. It
displays a panel of 4 diagnostic plots: top left - study region with example 
population, top right - study region with transects, bottom left - an example
realisation of a survey with detected animals shown in cyan and undetected
animals shown in red, bottom right - a histogram of the example distances to 
which the detection function would be fitted.
}
\examples{
\dontrun{
sim <- make.simulation(design.obj = make.design("point"))
check.sim.setup(sim)
}
}
\author{
Eric Rexstad, Laura Marshall
}

