% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{create.sample.table}
\alias{Survey-method}
\alias{create.sample.table}
\alias{create.sample.table,}
\alias{create.sample.table,Survey-method}
\title{S4 generic method to generate a sample table}
\usage{
create.sample.table(object)

\S4method{create.sample.table}{Survey}(object)
}
\arguments{
\item{object}{an object of a class inheriting from Survey}
}
\value{
an object of class Sample.Table
}
\description{
This function is called internally to generates a sample table required to
estimate abundance /  density via the Hortvitz-Thompson estimator.
}

