\docType{class}
\name{DDF.Analysis-class}
\alias{DDF.Analysis-class}
\title{Class "DDF.Analysis"}
\description{
  Class \code{"DDF.Analysis"} is an S4 class describing a
  detection function which is to be fitted to the data.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.ddf.analysis(dsmodel = ~cds(key = "hn",
  formula = ~1), criteria = "AIC")}
}
\keyword{classes}

