\docType{class}
\name{DDF.Data-class}
\alias{DDF.Data-class}
\title{Class "DDF.Data"}
\description{
  A virtual class containing a data.frame with distance
  sampling data in the correct format for mrds.
}
\section{Slots}{
  \describe{ \item{\code{ddf.dat}}{Object of class
  \code{"data.frame"}; dataframe with all the necessary
  column to fit a detection function using mrds.} }
}
\keyword{classes}

