% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/description.summary.R
\name{description.summary}
\alias{description.summary}
\title{Provides a description of the summary object/output}
\usage{
description.summary()
}
\description{
Prints a list of the terms used in the simulation summary.
}
\author{
Laura Marshall
}

