% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.Summary.R
\docType{class}
\name{Design.Summary-class}
\alias{Design.Summary-class}
\title{S4 Class "Design.Summary"}
\description{
Class "Design.Summary"
}
\details{
Class \code{"Design.Summary"} is an S4 class containing a summary of
the survey design. This is returned when \code{summary(Design)}
is called. If it is not assigned to a variable the object will be 
displayed via the \code{show} method.
}
\section{Methods}{

\describe{
 \item{\code{show}}{\code{signature=(object = "Design.Summary")}: prints
 the contents of the object in a user friendly format.}
 }
}
\keyword{classes}

