\docType{class}
\name{Detectability-class}
\alias{Detectability-class}
\title{Class "Detectability"}
\description{
  Class \code{"Detectability"} is an S4 class describing
  the probablity of detecting individuals / clusters in a
  population.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.detectability(key.function, scale.param,
  perp.truncation, rad.truncation)}
}
\keyword{classes}

