% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{get.area}
\alias{get.area}
\alias{get.area,Region-method}
\title{Returns the area of the region}
\usage{
get.area(object)

\S4method{get.area}{Region}(object)
}
\arguments{
\item{object}{object of class Region}
}
\value{
numeric value specifying the area of the region
}
\description{
Returns the area of the region
}

