\docType{methods}
\name{get.criteria}
\alias{get.criteria}
\title{S4 generic get method}
\arguments{
  \item{object}{an object of class simulation or
  DDF.Analysis}

  \item{object}{an object of class Simulation}

  \item{dht.tables}{logical value indicating whether or the
  data tables for Hortvitz-Thompson estimation are
  required.}

  \item{object}{an object of class Simulation or
  DDF.Analysis}
}
\value{
  the name of the criteria used for model selection

  an object of class Region.Table

  the model selection criteria value
}
\description{
  Accesses the information on the selection criteria

  Simulates the process by which individuals / clusters are
  detected.
}

