\docType{methods}
\name{get.distance.data}
\alias{get.distance.data}
\alias{get.distance.data,LT.Survey.Results-method}
\title{S4 generic method to extract distance data}
\arguments{
  \item{object}{an object of class LT.Survey.Results}
}
\value{
  a data.frame describing the distance data
}
\description{
  Extracts distance data from a Survey.Results object
}
\seealso{
  \code{\link{create.survey.results}}
}

