% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.Description.R
\docType{methods}
\name{get.N}
\alias{get.N}
\alias{get.N,Population.Description-method}
\title{S4 generic method to return N}
\usage{
get.N(object)

\S4method{get.N}{Population.Description}(object)
}
\arguments{
\item{object}{an object of class Population.Description}
}
\value{
numeric value of the population size
}
\description{
Returns the population size
}

