% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\docType{methods}
\name{hist,Simulation-method}
\alias{hist,Simulation-method}
\title{hist}
\usage{
\S4method{hist}{Simulation}(x, ...)
}
\arguments{
\item{object}{object of class Simulation}
}
\description{
Not currently implemented
}

