% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Line.Transect.R
\docType{class}
\name{Line.Transect-class}
\alias{Line.Transect-class}
\title{S4 Class "Line.Transect"}
\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "Line.Transect")}: plots the 
 transects.}
}
}
\keyword{classes}

