\docType{class}
\name{LT.Design-class}
\alias{LT.Design-class}
\title{Virtual Class "LT.Design"}
\description{
  Virtual Class \code{"LT.Design"} is an S4 class detailing
  the type of line transect design and the co-ordinates of
  the end points of the transects.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
}
\keyword{classes}

