\docType{class}
\name{LT.Survey-class}
\alias{LT.Survey-class}
\title{Virtual Class "LT.Survey"}
\description{
  Virtual Class \code{"LT.Survey"} is an S4 class
  containing a population and a set of transects.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
}
\keyword{classes}

