% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LT.Survey.Results.R
\docType{class}
\name{LT.Survey.Results-class}
\alias{LT.Survey.Results-class}
\title{S4 Class "LT.Survey.Results"}
\description{
S4 Class "LT.Survey.Results"
}
\section{Slots}{

\describe{
\item{\code{region}}{Object of class \code{"Region"}; the region
representation.}

\item{\code{population}}{Object of class \code{"Population"}; the
population.}

\item{\code{transects}}{Object of class \code{"Line.Transect"}; the
transects.}

\item{\code{ddf.data}}{Object of class \code{"Single.Obs.DDF.Data"}; The
ddf data for \code{ddf}.}

\item{\code{obs.table}}{Object of class \code{"Obs.Table"}; One of the
tables for \code{dht}.}

\item{\code{sample.table}}{Object of class \code{"Sample.Table"}; One of
the tables for \code{dht}.}

\item{\code{region.table}}{Object of class \code{"Region.Table"}; One of
the tables for \code{dht}.}
}}
\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "LT.Survey.Results")}: plots
 the region, the location of individuals in the population, the transects
 and the successful sightings.}
 \item{\code{get.distance.data}}{\code{signature=(object = "LT.Survey.Results")}: returns the ddf data as a dataframe..}
}
}
\keyword{classes}

