% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Obs.Table-class}
\alias{Obs.Table-class}
\title{S4 Class "Obs.Table"}
\description{
Class "Obs.Table"
}
\details{
Class \code{"Obs.Table"} is an S4 class containing an
observation table which is required for Hortvitz-Thompson estimation of
density and abundance.
}
\section{Objects from the Class}{
 Objects can be created by calls to the
function \code{create.survey.results(simulation, dht.table = TRUE)}
}
\seealso{
\code{\link{create.survey.results}}
}
\keyword{classes}

