\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,Density-method}
\alias{plot,Line.Transect-method}
\alias{plot,LT.Survey.Results-method}
\alias{plot,Population-method}
\alias{plot,Region-method}
\alias{plot,Simulation-method}
\alias{plot,Single.Obs.DDF.Data-method}
\title{S4 generic method to plot an  object}
\description{
  S4 generic method to plot an object
}
\author{
  Laura Marshall
}

