% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDF.Data.R
\docType{methods}
\name{plot,DDF.Data,ANY-method}
\alias{plot,DDF.Data,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{DDF.Data,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class DDF.Data}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'DDF.Data'. Requires that the
associated region has already been plotted. This function adds 
the locations of the individuals/clusters in the population who
were detected.
}

