% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\docType{methods}
\name{plot,Density,ANY-method}
\alias{plot,Density,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Density,ANY}(x, y, add = FALSE, plot.units = character(0),
  contours = TRUE, style = "points", density.col = heat.colors(12), ...)
}
\arguments{
\item{x}{object of class Density}

\item{y}{not used}

\item{add}{logical indicating whether it should be added to 
existing plot}

\item{plot.units}{allows for units to be converted between m
and km}

\item{contours}{logical indicating whether contours should be 
added}

\item{style}{character "points" or "blocks". Points displays
a coloured point at the centre of each grid cell where as 
blocks colours the entire cell.}

\item{density.col}{the colours used to indicate density level}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Density'
}

