% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Line.Transect.R
\docType{methods}
\name{plot,Line.Transect-method}
\alias{plot,Line.Transect-method}
\title{Plot}
\usage{
\S4method{plot}{Line.Transect}(x, y, transect.ID = numeric(0), col = 1, ...)
}
\arguments{
\item{x}{object of class Line.Transect}

\item{y}{not used}

\item{transect.ID}{allows individual or groups of transects
to be added}

\item{col}{colour of the lines}

\item{...}{other general plot parameters e.g. lwd}
}
\description{
Plots an S4 object of class 'Line.Transect'. Requires that the
associated region has already been plotted. This function adds
the transect lines.
}

