% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT.Survey.Results.R
\name{plot,LT.Survey.Results,ANY-method}
\alias{plot,LT.Survey.Results,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{LT.Survey.Results,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class LT.Survey.Results}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'LT.Survey.Results'. Plots the
region, the transects, the population and colour codes the 
detections
}
