% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.Transect.R
\docType{methods}
\name{plot,Point.Transect,ANY-method}
\alias{plot,Point.Transect,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Point.Transect,ANY}(x, y, transect.ID = numeric(0),
  col = 1, ...)
}
\arguments{
\item{x}{object of class Point.Transect}

\item{y}{not used}

\item{transect.ID}{allows individual or groups of transects 
to be added}

\item{col}{colour of the lines}

\item{...}{other general plot parameters e.g. lwd}
}
\description{
Plots an S4 object of class 'Point.Transect'. Requires that the
associated region has already been plotted. This function adds 
the transect lines.
}
