% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Population.R
\docType{methods}
\name{plot,Population-method}
\alias{plot,Population-method}
\title{Plot}
\usage{
\S4method{plot}{Population}(x, y, ...)
}
\arguments{
\item{x}{object of class Population}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Population'. Requires that the
associated region has already been plotted. This function adds
the locations of the individuals/clusters in the population.
}

