% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\docType{methods}
\name{plot,Region,ANY-method}
\alias{plot,Region,ANY-method}
\title{Plot}
\usage{
\S4method{plot}{Region,ANY}(x, y, add = FALSE, plot.units = character(0),
  region.col = NULL, gap.col = NULL, main = "", ...)
}
\arguments{
\item{x}{object of class Region}

\item{y}{not used}

\item{add}{logical indicating whether it should be added to 
existing plot}

\item{plot.units}{allows for units to be converted between m
and km}

\item{region.col}{fill colour for the region}

\item{gap.col}{fill colour for the gaps}

\item{main}{character plot title}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Region'
}
