% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.Results.R
\docType{methods}
\name{plot,Survey.Results-method}
\alias{plot,Survey.Results-method}
\title{Plot}
\usage{
\S4method{plot}{Survey.Results}(x, y, ...)
}
\arguments{
\item{x}{object of class LT.Survey.Results}

\item{y}{not used}

\item{...}{other general plot parameters}
}
\description{
Plots an S4 object of class 'Survey.Results'. Plots the
region, the transects, the population and colour codes the 
detections
}

