% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point.Transect.R
\docType{class}
\name{Point.Transect-class}
\alias{Point.Transect-class}
\title{S4 Class "Point.Transect"}
\description{
Class "Point.Transect" contains an instance of a Line Transect Survey
}
\section{Slots}{

\describe{
\item{\code{design.obj}}{Object of class \code{"character"}; the object name
of the design object which generated the transects.}

\item{\code{sampler.info}}{Object of class \code{"data.frame"}; the sampler point coordinates.}
}}
\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "Point.Transect")}: plots the 
 transects.}
}
}
\keyword{classes}

