\docType{class}
\name{Population-class}
\alias{plot,Population-method}
\alias{Population-class}
\title{Class "Population"}
\description{
  Contains an instance of a population including a
  description of their detectability in the form of an
  object of calss Detectability.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
}
\keyword{classes}

