\docType{class}
\name{Population.Description-class}
\alias{Population.Description-class}
\title{Class "Population.Description"}
\description{
  Class \code{"Population.Description"} is an S4 class
  containing a description of the population. It provides
  methods to generate an example population.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.population.description(N, density.obj,
  region)}
}
\keyword{classes}

