% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.Summary.R
\docType{class}
\name{Population.Summary-class}
\alias{Population.Summary-class}
\title{S4 Class "Population.Summary"}
\description{
Class "Population.Summary"
}
\details{
Class \code{"Population.Summary"} is an S4 class containing a summary of
the survey population. This is returned when \code{summary(Population)}
is called. If it is not assigned to a variable the object will be 
displayed via the \code{show} method.
}
\section{Methods}{

\describe{
 \item{\code{show}}{\code{signature=(object = "Population.Summary")}: prints
 the contents of the object in a user friendly format.}
 }
}
\keyword{classes}

