\docType{class}
\name{PT.Design-class}
\alias{PT.Design-class}
\title{Virtual Class "PT.Design"}
\description{
  Virtual Class \code{"PT.Design"} is an S4 class detailing
  the type of point transect design and the co-ordinates of
  the end points of the transects.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
}
\keyword{classes}

