\docType{class}
\name{Region-class}
\alias{Region-class}
\title{Class "Region"}
\description{
  Class \code{"Region"} is an S4 class containing
  descriptions of the study area. The polygons describing
  the region are found in the coords slot and any gaps are
  described as polygons in the gaps slot.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.region(region.name = "region.name", shapefile
  = region.shapefile)}
}
\keyword{classes}

