% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.duplicates.R
\name{rename.duplicates}
\alias{rename.duplicates}
\title{Renumbers the object IDs for the duplicate observations generated when
bootstrapping}
\usage{
rename.duplicates(ddf.dat, double.observer = FALSE)
}
\arguments{
\item{ddf.dat}{dataframe containing a single dataset with duplicate 
observations}

\item{double.observer}{logical indicating whether it is a double observer survey or not}
}
\value{
list with 2 elements:
  ddf.dat dataframe containing a single dataset with new and unique 
    observation IDs 
  obs.table the updated obs.table dataframe containing the new
    observation IDs
}
\description{
Find the largest object ID and renumbers all duplicate IDs starting form 
this value. The information for the duplicates is also added to the obs.table
}
\note{
Internal function not intended to be called by user.
}
\seealso{
\code{resample.data}
}
\author{
Laura Marshall
}
\keyword{data}
\keyword{manipulation}
