\docType{class}
\name{Sample.Table-class}
\alias{Sample.Table-class}
\title{Class "Sample.Table"}
\description{
  Class \code{"Sample.Table"} is an S4 class containing a
  region table which is required for Hortvitz-Thompson
  estimation of density and abundance.
}
\section{Objects from the Class}{
  Objects can be created by calls to the function
  \code{simulate.survey(simulation, dht.table = TRUE)}
}
\keyword{classes}

