% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.sim.results.R
\name{save.sim.results}
\alias{save.sim.results}
\title{save.sim.results}
\usage{
save.sim.results(simulation, filepath = character(0), sim.ID = numeric(0))
}
\arguments{
\item{simulation}{object of class \code{Simulation} which has been run.}

\item{filepath}{optionally a path to the directory where you would like the files saved, otherwise it will save it to the working directory.}

\item{sim.ID}{optionally you can add a simulation ID to the filename}
}
\value{
invisibly returns the original simulation object
}
\description{
Saves the simulation results from each replicate to file. It will save up to 3 csv files, one for the abundance estimation for individuals, one for the abundance estimation of clusters (where applicable) and one for detectability estimates and model selection information.
}
\author{
L. Marshall
}
