% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setcov.R
\name{setcov}
\alias{setcov}
\title{dmat: data matrix
model: model formula
returns a design matrix for the specified data and model
author: Jeff Laake}
\usage{
setcov(dmat, model)
}
\description{
dmat: data matrix
model: model formula
returns a design matrix for the specified data and model
author: Jeff Laake
}

