% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{show,Simulation-method}
\alias{show,Simulation-method}
\title{show}
\usage{
\S4method{show}{Simulation}(object)
}
\arguments{
\item{object}{object of class Simulation}
}
\description{
Not currently implemented
}
