% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulation.Summary.R
\docType{methods}
\name{show,Simulation.Summary-method}
\alias{show,Simulation.Summary-method}
\title{show}
\usage{
\S4method{show}{Simulation.Summary}(object)
}
\arguments{
\item{object}{object of class Simulation.Summary}
}
\description{
Displays the simulation summary
}

