\docType{class}
\name{Simulation.Summary-class}
\alias{show}
\alias{show,Simulation.Summary-method}
\alias{Simulation.Summary-class}
\title{Class "Simulation.Summary"}
\description{
  Class \code{"Simulation.Summary"} is an S4 class
  containing a summary of the simulation results
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.simulation(reps, double.observer, region.obj,
  design.obj, population.description.obj,
  detectability.obj, ddf.analyses.list)}
}
\keyword{classes}

