% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Single.Obs.LT.Survey.R
\docType{class}
\name{Single.Obs.LT.Survey-class}
\alias{Single.Obs.LT.Survey-class}
\title{S4 Class "Single.Obs.LT.Survey"}
\description{
S4 Class "Single.Obs.LT.Survey"
}
\keyword{classes}

