% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single.Obs.PT.Survey.R
\docType{class}
\name{Single.Obs.PT.Survey-class}
\alias{Single.Obs.PT.Survey-class}
\title{Class "Single.Obs.PT.Survey"}
\description{
An S4 class containing an instance of a population
and a set of transects.
}
\keyword{classes}
