\docType{methods}
\name{summary}
\alias{summary}
\title{S4 generic method to displays a summary of a Simulation object}
\arguments{
  \item{object}{an object of class simulation}
}
\description{
  If the simulation has been run this will include a
  summary of the results.
}
\author{
  Laura Marshall
}

