% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{summary,Simulation-method}
\alias{summary,Simulation-method}
\title{summary}
\usage{
\S4method{summary}{Simulation}(object, description.summary = TRUE, ...)
}
\arguments{
\item{object}{object of class Simulation}

\item{description.summary}{logical indicating whether an
explanation of the summary should be included}

\item{...}{can specify if you want the maximum number of iterations to be used where at least one model converged (use.max.reps = TRUE) or only use iterations where all models converged (use.max.reps = FALSE)}
}
\description{
Provides a summary of the simulation results.
}
