\docType{class}
\name{Survey-class}
\alias{Survey-class}
\title{Class "Survey"}
\description{
  Class \code{"Survey"} is an S4 class containing an
  instance of a population and a set of transects.
}
\section{Objects from the Class}{
  Objects which inherit from this class are usually created
  internally, however, an instance is also contained in the
  object returned from \code{simulate.survey(simulation)}
}
\keyword{classes}

