\docType{class}
\name{Survey.Design-class}
\alias{Survey.Design-class}
\title{Virtual Class "Survey.Design"}
\description{
  Virtual Class \code{"Survey.Design"} is an S4 class
  detailing the type of line transect design and the
  co-ordinates of the end points of the transects.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.design(transect.type, design.details, region,
  design.axis, spacing, plus.sampling, path)}
}
\keyword{classes}

