% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transect.R
\docType{class}
\name{Transect-class}
\alias{Transect-class}
\title{S4 Virtual Class "Transect"}
\description{
Class "Transect" contains an instance of a Line Transect Survey
}
\section{Slots}{

\describe{
\item{\code{design.obj}}{Object of class \code{"character"}; the object name
of the design object which generated the transects.}

\item{\code{sampler.info}}{Object of class \code{"data.frame"}; the sampler 
coordinates.}
}}

\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "Transect")}: plots the 
 transects.}
}
}

\keyword{classes}
