\docType{methods}
\name{create.sample.table}
\alias{create.sample.table}
\alias{create.sample.table,}
\alias{Survey-method}
\title{S4 generic method to generate a sample table}
\arguments{
  \item{object}{an object of a class inheriting from
  Survey}
}
\value{
  an object of class Sample.Table
}
\description{
  This function is called internally to generates a sample
  table required to estimate abundance / density via the
  Hortvitz-Thompson estimator.
}

