% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.for.distance.R
\name{data.for.distance}
\alias{data.for.distance}
\title{data.for.distance}
\usage{
data.for.distance(object, file = NULL, round = 2, sep = "\\t")
}
\arguments{
\item{object}{an object of class Survey.Results generated by create.survey.results}

\item{file}{path and filename if the user would like the results saved to file. Should contain the .txt file extension.}

\item{round}{the amount of decimal places to round the distances to}

\item{sep}{the field separator string for writing to file}
}
\description{
Formats the data generated by create.survey.results into
a form suitable for import into Distance
}
\author{
L Marshall
}
