% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.for.distance.R
\name{data.for.distance}
\alias{data.for.distance}
\title{data.for.distance}
\usage{
data.for.distance(object, file = NULL)
}
\arguments{
\item{object}{an object of class Survey.Results generated by create.survey.results}

\item{file}{if provided the data will be written to file. Should be a
filename ending in .csv}
}
\description{
Formats the data generated by create.survey.results into
a form suitable for import into Distance
}
\author{
L Marshall
}

