% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{DDF.Analysis-class}
\alias{DDF.Analysis-class}
\alias{run.analysis,DDF.Analysis,DDF.Data-method}
\title{S4 Class "DDF.Analysis"}
\description{
S4 Class "DDF.Analysis"
}
\section{Slots}{

\describe{
 \item{\code{dsmodel}}{Object of class \code{"formula"}; describing the
 detection function model.}
 \item{\code{criteria}}{Object of class \code{"character"}; describes
 which model selection criteria to use ("AIC","AICc","BIC").}
 \item{\code{truncation}}{Object of class \code{"character"}; Specifies 
 the truncation distance for the analyses.}
 \item{\code{binned.data}}{Object of class \code{"character"}; logical 
 value specifying if the data should be binned for analysis.}
 \item{\code{cutpoints}}{Object of class \code{"character"}; gives the 
 cutpoints of the bins for binned data analysis.}
 \item{\code{ddf.result}}{Object of class \code{"list"}; object of S3 class
 ddf.}
}
}
\section{Methods}{

\describe{
 \item{\code{run.analysis}}{\code{signature=c(object = "DDF.Analysis",
 data = "DDF.Data")}: runs the analysis described in the object on the
 data provided.}
}
}
\seealso{
\code{\link{make.ddf.analysis.list}}
}
\keyword{classes}

