% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDF.Analysis.R
\docType{class}
\name{DDF.Analysis-class}
\alias{DDF.Analysis-class}
\title{Class "DDF.Analysis"}
\description{
Class \code{"DDF.Analysis"} is an S4 class describing a detection function
which is to be fitted to the data.
}
\section{Slots}{

\describe{
\item{\code{dsmodel}}{Object of class \code{"formula"}; describing the
detection function model.}

\item{\code{criteria}}{Object of class \code{"character"}; describes 
which model selection criteria to use ("AIC","AICc","BIC").}

\item{\code{truncation}}{Object of class \code{"character"}; Specifies 
the truncation distance for the analyses.}

\item{\code{binned.data}}{Object of class \code{"character"}; logical 
value specifying if the data should be binned for analysis.}

\item{\code{cutpoints}}{Object of class \code{"character"}; gives the 
cutpoints of the bins for binned data analysis.}

\item{\code{analysis.strata}}{Dataframe with two columns ("design.id" and 
"analysis.id"). The former gives the strata names as defined in the
design (i.e. the region object) the second specifies how they should 
be grouped (into less strata) for the analyses}

\item{\code{ddf.result}}{Object of class \code{"list"}; object of S3 class
ddf.}
}}

\section{Methods}{

\describe{
 \item{\code{run.analysis}}{\code{signature=c(object = "DDF.Analysis", 
 data = "DDF.Data")}: runs the analysis described in the object on the
 data provided.}
}
}

\seealso{
\code{\link{make.ddf.analysis.list}}
}
\keyword{classes}
