\docType{class}
\name{Density-class}
\alias{Density-class}
\alias{plot,Density-method}
\title{Class "Density"}
\description{
  Class \code{"Density"} is an S4 class containing a grid
  which described the density of individuals / clusters of
  a population.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{make.density(region, x.space, y.space, constant)}
}
\keyword{classes}

