% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\docType{class}
\name{Density-class}
\alias{Density-class}
\title{Class "Density"}
\description{
Class \code{"Density"} is an S4 class containing a list of grids which
describe the density of individuals / clusters of a population. The list
contains one grid (\code{data.frame}) for each strata.
}
\section{Slots}{

\describe{
\item{\code{region.name}}{Object of class \code{"character"}; the region name.}

\item{\code{strata.name}}{Object of class \code{"character"}; the strata names}

\item{\code{density.surface}}{Object of class \code{"list"}; list of data.frames
with the columns x, y and density. There must be one data.frame for each
strata.}

\item{\code{x.space}}{Object of class \code{"numeric"}; The spacing between
gridpoints described in the density data.frames in the x-direction.}

\item{\code{y.space}}{Object of class \code{"numeric"}; The spacing between
gridpoints described in the density data.frames in the y-direction.}

\item{\code{units}}{Object of class \code{"numeric"}; The units of the grid 
points.}
}}

\seealso{
\code{\link{make.density}}
}
\keyword{classes}
