% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.summary.R
\name{description.summary}
\alias{description.summary}
\title{Provides a description of the summary object/output  
 
Prints a list of the terms used in the simulation summary.}
\usage{
description.summary()
}
\description{
Provides a description of the summary object/output  
 
Prints a list of the terms used in the simulation summary.
}
\author{
Laura Marshall
}
