% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Detectability-class}
\alias{Detectability-class}
\title{S4 Class "Detectability"}
\description{
Class "Detectability"
}
\details{
Class \code{"Detectability"} is an S4 class describing the probablity
of detecting individuals / clusters in a population.
}
\section{Slots}{

\describe{
 \item{\code{key.function}}{Object of class \code{"character"}; a code
 specifying the detection function form ("hn" = half normal, "hr" =
 hazard rate.)}
 \item{\code{scale.param}}{Object of class \code{"numeric"}; The scale
 parameter for the detection function.}
 \item{\code{shape.param}}{Object of class \code{"numeric"}; The shape
 parameter for the detection function.}
 \item{\code{covariates}}{Object of class \code{"character"}; The names
 of the covariates which affect detectability. Not yet implemented}
 \item{\code{cov.param}}{Object of class \code{"numeric"}; The parameter
 values associated with the covariates. Not yet implemented}
 \item{\code{truncation}}{Object of class \code{"numeric"}; The maximum
 distance at which objects may be detected.}
}
}
\seealso{
\code{\link{make.detectability}}
}
\keyword{classes}

