\docType{methods}
\name{get.criteria}
\alias{get.criteria}
\title{S4 generic get method}
\arguments{
  \item{object}{an object of class simulation or
  DDF.Analysis}

  \item{object}{an object of class Simulation or
  DDF.Analysis}
}
\value{
  the name of the criteria used for model selection

  the model selection criteria value
}
\description{
  Accesses the information on the selection criteria

  Provides the selection criteria value
}

