% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/LT.Survey.Results.R,
%   R/Survey.Results.R
\docType{methods}
\name{get.distance.data}
\alias{get.distance.data}
\alias{get.distance.data,LT.Survey.Results-method}
\alias{get.distance.data,Survey.Results-method}
\title{S4 generic method to extract distance data}
\usage{
get.distance.data(object)

\S4method{get.distance.data}{LT.Survey.Results}(object)

\S4method{get.distance.data}{Survey.Results}(object)
}
\arguments{
\item{object}{an object of class LT.Survey.Results}
}
\value{
a data.frame describing the distance data
}
\description{
Extracts distance data from a Survey.Results object
}
\seealso{
\code{\link{create.survey.results}}
}
