% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LT.Survey.Results.R, R/generic.functions.R
\docType{methods}
\name{get.distance.data,LT.Survey.Results-method}
\alias{get.distance.data}
\alias{get.distance.data,LT.Survey.Results-method}
\title{S4 generic method to extract distance data}
\usage{
\S4method{get.distance.data}{LT.Survey.Results}(object)

get.distance.data(object)
}
\arguments{
\item{object}{an object of class LT.Survey.Results}
}
\value{
a data.frame describing the distance data
}
\description{
Extracts distance data from a Survey.Results object
}
\seealso{
\code{\link{create.survey.results}}
}

