\docType{class}
\name{Line.Transect-class}
\alias{Line.Transect-class}
\alias{lmList-class}
\alias{show,lmList-method}
\title{Class "Line.Transect" containing an instance of a Line Transect Survey}
\description{
  Class \code{"LTSurvey"} is an S4 class detailing a set of
  line transect
}
\section{Objects from the Class}{

}
\keyword{classes}

