\docType{class}
\name{Line.Transect-class}
\alias{Line.Transect-class}
\alias{plot}
\alias{plot,Line.Transect-method}
\title{Class "Line.Transect" containing an instance of a Line Transect Survey}
\description{
  Class \code{"LTSurvey"} is an S4 class detailing a set of
  line transect
}
\keyword{classes}

