% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Line.Transect.R
\docType{class}
\name{Line.Transect-class}
\alias{Line.Transect-class}
\title{S4 Class "Line.Transect"}
\description{
Class "Line.Transect" contains an instance of a Line Transect Survey
}
\section{Slots}{

\describe{
\item{\code{design.obj}}{Object of class \code{"character"}; the object name
of the design object which generated the transects.}

\item{\code{sampler.info}}{Object of class \code{"data.frame"}; the sampler
end point coordinates.}
}}
\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "Line.Transect")}: plots the
 transects.}
}
}
\keyword{classes}

