% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{LT.Design-class}
\alias{LT.Design-class}
\title{S4 Class "LT.Design"}
\description{
S4 Class "LT.Design"
}
\section{Slots}{

\describe{
 \item{\code{design.axis}}{Object of class \code{"numeric"}; the angle of
 the design axis.}
 \item{\code{spacing}}{Object of class \code{"numeric"}; the spacing of
 the design.}
 }
}

\section{Methods}{

\describe{
 \item{\code{generate.transects}}{\code{signature=(object = "LT.Design", ...)}:
 loads a set of transects from a shapefile.}
}
}
\seealso{
\code{\link{make.design}}
}
\keyword{classes}

