% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT.Design.R
\docType{class}
\name{LT.Design-class}
\alias{LT.Design-class}
\title{Virtual Class "LT.Design" extends Class "Survey.Design"}
\description{
Virtual Class \code{"LT.Design"} is an S4 class detailing the type of 
line transect design.
}
\section{Methods}{

\describe{
 \item{\code{generate.transects}}{\code{signature=(object = "LT.Design", ...)}:
 loads a set of transects from a shapefile.}
}
}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
