% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{LT.Survey-class}
\alias{LT.Survey-class}
\title{S4 Class "LT.Survey"}
\description{
S4 Class "LT.Survey"
}
\section{Slots}{

\describe{
 \item{\code{line.transect}}{Object of class \code{"Line.Transect"}; the
 transects.}
 \item{\code{perpendicular.truncation}}{Object of class \code{"numeric"}; the
 maximum distance from the transect at which animals may be detected.}
 }
}

\section{Methods}{

\describe{
 \item{\code{create.sample.table}}{\code{signature=(object = "LT.Survey", ...)}:
 creates a sample table for \code{dht}.}
}
}
\seealso{
\code{\link{make.design}}
}
\keyword{classes}

