% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT.Survey.R
\docType{class}
\name{LT.Survey-class}
\alias{LT.Survey-class}
\title{Virtual Class "LT.Survey" extends class "Survey"}
\description{
Virtual Class \code{"LT.Survey"} is an S4 class containing a population
and a set of transects.
}
\section{Slots}{

\describe{
\item{\code{perpendicular.truncation}}{Object of class \code{"numeric"}; the
maximum distance from the transect at which animals may be detected.}
}}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
