% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{LT.Survey.Results-class}
\alias{LT.Survey.Results-class}
\alias{plot,LT.Survey.Results,ANY-method}
\alias{plot,LT.Survey.Results-method}
\title{S4 Class "LT.Survey.Results"}
\usage{
\S4method{plot}{LT.Survey.Results,ANY}(x, y, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }
}
\description{
S4 Class "LT.Survey.Results"
}
\section{Slots}{

\describe{
 \item{\code{region}}{Object of class \code{"Region"}; the region
 representation.}
 \item{\code{population}}{Object of class \code{"Population"}; the
 population.}
 \item{\code{transects}}{Object of class \code{"Line.Transect"}; the
 transects.}
 \item{\code{ddf.data}}{Object of class \code{"Single.Obs.DDF.Data"}; The
 ddf data for \code{ddf}.}
 \item{\code{obs.table}}{Object of class \code{"Obs.Table"}; One of the
 tables for \code{dht}.}
 \item{\code{sample.table}}{Object of class \code{"Sample.Table"}; One of
 the tables for \code{dht}.}
 \item{\code{region.table}}{Object of class \code{"Region.Table"}; One of
 the tables for \code{dht}.}
}
}

\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "LT.Survey.Results")}: plots
 the region, the location of individuals in the population, the transects
 and the successful sightings.}
 \item{\code{get.distance.data}}{\code{signature=(object = "LT.Survey.Results")}: returns the ddf data as a dataframe..}
}
}
\keyword{classes}

