% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.detectability}
\alias{make.detectability}
\title{Creates a Detectablility object}
\usage{
make.detectability(key.function = "hn", scale.param = 25,
  shape.param = numeric(0), cov.param = list(), truncation = 75)
}
\arguments{
\item{key.function}{specifies shape of the detection function (either 
half-normal "hn" or hazard rate "hr")}

\item{scale.param}{parameter value associated with the detection function}

\item{shape.param}{parameter value associated with the detection function}

\item{cov.param}{named list with parameters for covariates}

\item{truncation}{the maximum perpendicular (or radial) distance at which 
objects may be detected from a line (or point) transect.}
}
\value{
object of class Detectablility
}
\description{
The detectability of the population is described by the values in this 
class.
}
\examples{
# The default values create a detectability object as follows:
detect <- make.detectability(key.function = "hn", scale.param = 25,
 truncation = 75) 
}
\author{
Laura Marshall
}

