% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.Constructors.R
\name{make.detectability}
\alias{make.detectability}
\title{Creates a Detectablility object}
\usage{
make.detectability(key.function, scale.param, shape.param = numeric(0),
  covariates = character(0), cov.param = numeric(0), truncation)
}
\arguments{
\item{key.function}{specifies shape of the detection function (either 
half-normal "hn" or hazard rate "hr")}

\item{scale.param}{parameter value associated with the detection function}

\item{shape.param}{parameter value associated with the detection function}

\item{covariates}{not yet implemented}

\item{cov.param}{not yet implemented}

\item{truncation}{the maximum perpendicular (or radial) distance at which 
objects may be detected from a line (or point) transect.}
}
\value{
object of class Detectablility
}
\description{
The detectability of the population is described by the values in this 
class.
}
\examples{
detect <- make.detectability(key.function = "hn", scale.param = 15,
 truncation = 30) 
}
\author{
Laura Marshall
}

