% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Population-class}
\alias{Population-class}
\alias{plot,Population,ANY-method}
\alias{plot,Population-method}
\title{S4 Class "Population"}
\usage{
\S4method{plot}{Population,ANY}(x, y, type = "p")
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}
    
\item{type}{supplies the type for the Population plot method}
}
\description{
S4 Class "Population"
}
\section{Slots}{

\describe{
 \item{\code{region.obj}}{Object of class \code{"character"}; the name of the region
 object.}
 \item{\code{strata.names}}{Object of class \code{"character"}; the names of the
 strata.}
 \item{\code{N}}{Object of class \code{"numeric"}; the number of individuals/clusters.}
 \item{\code{D}}{Object of class \code{"numeric"}; the density of individuals/clusters.}
 \item{\code{population}}{Object of class \code{"data.frame"}; the locations of
 individuals/clusters and any population covariates.}
 \item{\code{detectability}}{Object of class \code{"Detectability"}; describes how
 easily the individuals/clusters can be detected.}
}
}

\section{Methods}{

\describe{
 \item{\code{plot}}{\code{signature=(object = "Line.Transect")}: plots the locations
 of the individuals/clusters.}
}
}
\seealso{
\code{\link{make.population.description}}, \code{\link{make.detectability}}
}
\keyword{classes}

