% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.Description.R
\docType{class}
\name{Population.Description-class}
\alias{Population.Description-class}
\title{Class "Population.Description"}
\description{
Class \code{"Population.Description"} is an S4 class containing a
description of the population. It provides methods to generate an
example population.
}
\section{Slots}{

\describe{
\item{\code{N}}{Object of class \code{"numeric"}; number of individuals
in the population (optional).}

\item{\code{density}}{Object of class \code{"Density"}; describes the
population density}

\item{\code{region.name}}{Object of class \code{"character"}; name of
the region in which the population exists.}

\item{\code{size}}{Object of class \code{"logical"}; Indicating whether
detections will be made on cluster (TRUE) or individuals (FALSE).}

\item{\code{size.table}}{Object of class \code{"data.frame"}; This must
have 2 columns- \code{size} and \code{prob}. The first column gives the
possible cluster sizes and the second describes the probabilities
of each cluster size.}

\item{\code{gen.by.N}}{Object of class \code{"logical"}; If \code{TRUE}
N is fixed otherwise it is generated from a Poisson distribution.}

\item{\code{D.dist}}{Object of class \code{character}; Describes the
density distribution (currently not implemented).}
}}
\section{Methods}{

\describe{
 \item{\code{get.N}}{\code{signature=(object = "Population.Description")}:
 returns the value of \code{N}}
 \item{\code{generate.population}}{\code{signature=(object = "Population.Description")}: generates a single realisation of the population.}
}
}
\seealso{
\code{\link{make.population.description}}
}
\keyword{classes}

