% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PT.Design.R
\docType{class}
\name{PT.Design-class}
\alias{PT.Design-class}
\title{Virtual Class "PT.Design" - NOT YET IMPLEMENTED}
\description{
Virtual Class \code{"PT.Design"} is an S4 class detailing the type of point
transect design and the co-ordinates of the end points of the transects.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
}
\keyword{classes}

