% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.Table.R
\docType{class}
\name{Region.Table-class}
\alias{Region.Table-class}
\title{S4 Class "Region.Table"}
\description{
Class "Region.Table"
}
\details{
Class \code{"Region.Table"} is an S4 class containing a 
region table which is required for Hortvitz-Thompson estimation of 
density and abundance.
}
\section{Slots}{

\describe{
\item{\code{region.table}}{data.frame which is a region.table for dht}
}}

\section{Objects from the Class}{
 Objects can be created by calls to the 
function \code{create.survey.results(simulation, dht.table = TRUE)}
}

\seealso{
\code{\link{create.survey.results}}
}
\keyword{classes}
