% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.Table.R
\docType{class}
\name{Sample.Table-class}
\alias{Sample.Table-class}
\title{Class "Sample.Table"}
\description{
Class \code{"Sample.Table"} is an S4 class containing a 
region table which is required for Hortvitz-Thompson estimation of 
density and abundance.
}
\section{Slots}{

\describe{
\item{\code{sample.table}}{data.frame which is the sample table for dht}
}}
\section{Objects from the Class}{
 Objects can be created by calls to the 
function \code{create.survey.results(simulation, dht.table = TRUE)}
}
\seealso{
\code{\link{create.survey.results}}
}
\keyword{classes}

