% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.Summary.R, R/Simulation.Summary.R
\name{show,Design.Summary-method}
\alias{show,Design.Summary-method}
\alias{show,Simulation.Summary-method}
\title{show}
\usage{
\S4method{show}{Design.Summary}(object)

\S4method{show}{Simulation.Summary}(object)
}
\arguments{
\item{object}{object of class Simulation.Summary}
}
\description{
Displays the simulation summary

Displays the simulation summary
}
\details{
#@param object object of class Simulation.Summary
}
