\docType{methods}
\name{simulate.survey}
\alias{simulate.survey}
\title{S4 generic method to simulate a survey}
\arguments{
  \item{object}{an object of class Simulation}

  \item{dht.tables}{logical value indicating whether or the
  data tables for Hortvitz-Thompson estimation are
  required.}
}
\value{
  an object of class Region.Table
}
\description{
  Simulates the process by which individuals / clusters are
  detected.
}

