% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Simulation.Summary-class}
\alias{Simulation.Summary-class}
\alias{show,Simulation.Summary-method}
\title{S4 Class "Simulation.Summary"}
\usage{
\S4method{show}{Simulation.Summary}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Class "Simulation.Summary"
}
\details{
Class \code{"Simulation.Summary"} is an S4 class containing a summary of
the simulation results. This is returned when \code{summary(Simulation)}
is called. If it is not assigned to a variable the object will be
displayed via the \code{show} method.
}
\section{Methods}{

\describe{
 \item{\code{show}}{\code{signature=(object = "Simulation.Summary")}: prints
 the contents of the object in a user friendly format.}
 }
}
\keyword{classes}

