\docType{class}
\name{Single.Obs.DDF.Data-class}
\alias{plot,Single.Obs.DDF.Data-method}
\alias{Single.Obs.DDF.Data-class}
\title{S4 Class "Single.Obs.DDF.Data"}
\description{
  DDF data resulting from a single observer survey.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
}
\keyword{classes}

