% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single.Obs.DDF.Data.R
\docType{class}
\name{Single.Obs.DDF.Data-class}
\alias{Single.Obs.DDF.Data-class}
\title{S4 Class "Single.Obs.DDF.Data"}
\description{
DDF data resulting from a single observer survey.
}
\keyword{classes}

