% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulation.R
\docType{methods}
\name{summary,Simulation-method}
\alias{summary,Simulation-method}
\title{summary}
\usage{
\S4method{summary}{Simulation}(object, description.summary = TRUE, ...)
}
\arguments{
\item{object}{object of class Simulation}

\item{description.summary}{logical indicating whether an
explanation of the summary should be included}

\item{...}{not implemented}
}
\description{
Provides a summary of the simulation results.
}

