\docType{class}
\name{Survey-class}
\alias{Survey-class}
\title{Virtual Class "Survey"}
\description{
  Class \code{"Survey"} is an S4 class containing an
  instance of a population.
}
\section{Slots}{
  \describe{ \item{\code{population}}{Object of class
  \code{"Population"}; an instance of a population.} }
}

\section{Methods}{
  \describe{
  \item{\code{create.region.table}}{\code{signature=(object
  = "Survey", ...)}: creates a region table for
  \code{dht}.} }
}
\keyword{classes}

