% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Survey.R
\docType{class}
\name{Survey-class}
\alias{Survey-class}
\title{S4 Class "Survey"}
\description{
S4 Class "Survey"
}
\section{Slots}{

\describe{
\item{\code{population}}{Object of class \code{"Population"}; an instance of
a population.}
}}
\section{Methods}{

\describe{
 \item{\code{create.region.table}}{\code{signature=(object = "Survey", ...)}:
 creates a region table for \code{dht}.}
}
}
\keyword{classes}

